clean up old code and move /NewDinners to /Dinners
