+++ src/pages/dinners.tsx
@@ -1,5 +1,5 @@
 import Head from "next/head";
-import { DinnersView } from "../views/NewDinners/DinnersView";
+import { DinnersView } from "../views/Dinners/DinnersView";
 
 export default function Home() {
   return (
