--- src/views/Dinners/Filter.tsx
@@ -1,3 +1,0 @@
-export const Filter = () => {
-  return <>Filter</>;
-};
