--- src/views/Dinners/Search.tsx
@@ -1,3 +1,0 @@
-export const Search = () => {
-  return <>Search</>;
-};
